Created various basic predefined obstacles
Still need to randomly generate them as the player progresses, work out a frequency and pattern that is challenging but not frustrating. As of now, the obstacles move from left to right, are unaffected by gravity or the player, will kill the player and reset level on collision. Will also destroy game object as the lava overtakes it.